kiwi_desktop/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-intake-workbench"
version = "0.1.0"
description = "Desktop tool for capturing and organizing knowledge intake (SQLite-backed)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Knowledge Intake Workbench Contributors" }]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
    "PySide6>=6.8.0",
    "python-pptx>=0.6.21",
    "python-docx>=1.1.0",
    "pymupdf>=1.24.0",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
kiw = "main:main"

[project.urls]
Homepage = "https://example.com/knowledge-intake-workbench"

[tool.setuptools]
packages = ["db", "db.repositories", "models", "services", "cli", "utils", "gui"]
py-modules = ["main"]

[tool.setuptools.package-data]
db = ["schema.sql"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
